07 hands on ml ageron

+++ hands_on_ml_ageron/05_exercises.py
@@ -77,8 +77,8 @@ def exercise10():
     y = dataset['target']
     X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2)
     scaler = StandardScaler()
-    X_train_scaled = scaler.fit_transform(X_train)
-    X_test_scaled = scaler.transform(X_test)
+    X_train_scaled = scaler.fit_transform(X_train.astype(np.float32))
+    X_test_scaled = scaler.transform(X_test.astype(np.float32))
 
     svr = SVR()
     params = {'gamma': reciprocal(0.001, 0.1), 'C': uniform(1, 10)}
